+++ CHANGELOG.md
@@ -4,6 +4,22 @@
 - Documentation for `AccountInfo`.
 - Fix bug in `AccountInfo` such that it supports `ScheduledRelease`s.
 
+
+## 1.6.1
+- Fix a bug which caused the `Client` to use a global absolute timeout from when it was initialized.
+
+## 1.6.0
+- `AccountTransaction` now exposes a `getType` function which
+   returns the type of the underlying `Payload`.
+
+## 1.5.0
+- Exposed more of the `transactions` in the API.
+
+## 1.4.0
+- Fix parsing bug for `AccountInfo` where scheduled releases was not included.
+- Transaction `Status` is now public.
+- `BlockItem` is now public.
+
 ## 1.3.0
 - Support for transaction deserialization.
 - Support for `getBlockInfo`.